requirements.txt
django
djangorestframework
rules
pyyaml
pytest
pytest-django
pytest-sugar
